run upgrade, extract translations, fix geoprocessing.json import paths
