+++ package.json
@@ -2,105 +2,103 @@
   "name": "fsm-nearshore-worker",
   "version": "6.1.2",
   "description": "FSM nearshore reports",
-  "private": false,
-  "type": "module",
-  "scripts": {
-    "start-data": "http-server data/dist -c-1",
-    "__test": "TEST_ROOT=$(pwd) geoprocessing test",
-    "test": "start-server-and-test start-data 8080 'npm run __test'",
-    "test:matching": "npm run __test -- -t",
-    "add:template": "geoprocessing add:template",
-    "import:data": "geoprocessing import:data",
-    "reimport:data": "geoprocessing reimport:data",
-    "precalc:data": "start-server-and-test 'http-server data/dist -c-1 -p 8001' http://localhost:8001 precalc:data_",
-    "precalc:data_": "geoprocessing precalc:data",
-    "precalc:data:clean": "geoprocessing precalc:data:clean",
-    "publish:data": "geoprocessing publish:data",
-    "install:scripts": "mkdir -p scripts && cp -r node_modules/@seasketch/geoprocessing/dist/base-project/scripts/* scripts",
-    "translation:install": "npx tsx scripts/translationInstall.ts",
-    "translation:extract": "npx tsx scripts/translationExtract.ts",
-    "translation:publish": "npx tsx src/i18n/bin/publishTerms.ts",
-    "translation:import": "npx tsx src/i18n/bin/importTerms.ts",
-    "translation:sync": "npm run translation:extract && npm run translation:publish && npm run translation:import",
-    "create:function": "geoprocessing create:function",
-    "create:client": "geoprocessing create:client",
-    "create:report": "geoprocessing create:report",
-    "start:client": "geoprocessing start:client",
-    "synth": "geoprocessing synth",
-    "bootstrap": "geoprocessing bootstrap",
-    "deploy": "geoprocessing deploy",
-    "destroy": "geoprocessing destroy",
-    "build": "geoprocessing build:lambda && geoprocessing build:client",
-    "build:client": "geoprocessing build:client",
-    "build:lambda": "geoprocessing build:lambda",
-    "storybook": "PROJECT_PATH=$(pwd) npx tsx .storybook/genReportStories.ts && storybook dev -p 6006",
-    "storybook:install": "npx tsx scripts/storybookInstall.ts",
-    "url": "geoprocessing url",
-    "clear-results": "geoprocessing clear-results",
-    "clear-all-results": "geoprocessing clear-all-results",
-    "prepare": "npm run translation:install"
+  "author": "Abby Vath Meyer",
+  "license": "BSD-3-Clause",
+  "homepage": "https://github.com/seasketch/fsm-nearshore-reports#readme",
+  "bugs": {
+    "url": "https://github.com/seasketch/fsm-nearshore-reports/issues"
   },
-  "keywords": [
-    "@seasketch/geoprocessing"
-  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/seasketch/fsm-nearshore-reports.git"
+  },
   "dependencies": {
-    "@turf/area": "7.0.0-alpha.113",
-    "@turf/bbox": "7.0.0-alpha.113",
-    "@turf/bbox-clip": "7.0.0-alpha.113",
-    "@turf/boolean-valid": "7.0.0-alpha.113",
-    "@turf/combine": "7.0.0-alpha.113",
-    "@turf/flatten": "7.0.0-alpha.113",
-    "@turf/helpers": "7.0.0-alpha.113",
-    "@turf/invariant": "7.0.0-alpha.113",
-    "@turf/kinks": "7.0.0-alpha.113",
-    "@turf/simplify": "7.0.0-alpha.113",
+    "@turf/area": "7.0.0",
+    "@turf/bbox": "7.0.0",
+    "@turf/bbox-clip": "7.0.0",
+    "@turf/boolean-valid": "7.0.0",
+    "@turf/combine": "7.0.0",
+    "@turf/flatten": "7.0.0",
+    "@turf/helpers": "7.0.0",
+    "@turf/invariant": "7.0.0",
+    "@turf/kinks": "7.0.0",
+    "@turf/simplify": "7.0.0",
     "@types/d3": "^7.4.3",
     "d3": "^7.9.0",
     "d3-scale": "^4.0.2",
-    "geoblaze": "2.6.0",
+    "geoblaze": "2.6.1",
     "geobuf": "^3.0.2",
     "geojson-antimeridian-cut": "^0.1.0",
     "pascalcase": "^1.0.0",
     "pbf": "^3.2.1",
-    "react-i18next": "^12.2.0",
+    "react-i18next": "^14.1.1",
     "react-to-print": "^2.15.1",
     "styled-components": "^6.1.8",
     "threads": "^1.7.0",
     "union-subdivided-polygons": "^0.9.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.24.1",
+    "@babel/cli": "^7.24.5",
     "@babel/plugin-syntax-import-attributes": "^7.24.1",
-    "@seasketch/geoprocessing": "7.0.0-beta.3",
-    "@storybook/addon-essentials": "^8.0.10",
-    "@storybook/addon-interactions": "^8.0.10",
-    "@storybook/addon-links": "^8.0.10",
-    "@storybook/blocks": "^8.0.10",
-    "@storybook/react": "^8.0.10",
-    "@storybook/react-vite": "^8.0.10",
-    "@storybook/test": "^8.0.10",
-    "@turf/boolean-disjoint": "7.0.0-alpha.113",
-    "@turf/boolean-valid": "7.0.0-alpha.113",
-    "@types/request": "^2.48.8",
+    "@seasketch/geoprocessing": "7.0.0-experimental-project-refactor.25",
+    "@storybook/addon-essentials": "^8.1.1",
+    "@storybook/addon-interactions": "^8.1.1",
+    "@storybook/addon-links": "^8.1.1",
+    "@storybook/blocks": "^8.1.1",
+    "@storybook/react": "^8.1.1",
+    "@storybook/react-vite": "^8.1.1",
+    "@storybook/test": "^8.1.1",
+    "@turf/boolean-disjoint": "7.0.0",
+    "@turf/boolean-valid": "7.0.0",
+    "@types/request": "^2.48.12",
     "aws-sdk-mock": "^6.0.1",
     "babel-plugin-i18next-extract": "0.10.0",
     "globby": "^14.0.1",
     "prop-types": "^15.8.1",
     "request": "^2.88.2",
-    "storybook": "^8.0.10",
+    "storybook": "^8.1.1",
     "vite": "^5.2.11",
-    "vitest": "^1.5.2",
-    "zx": "^8.0.2"
+    "vitest": "^1.6.0",
+    "zx": "^8.1.0"
   },
-  "repositoryUrl": "https://github.com/seasketch/fsm-nearshore-reports",
-  "author": "Abby Vath Meyer",
-  "license": "BSD-3-Clause",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/seasketch/fsm-nearshore-reports.git"
+  "scripts": {
+    "__test": "TEST_ROOT=$(pwd) geoprocessing test",
+    "add:template": "geoprocessing add:template",
+    "bootstrap": "geoprocessing bootstrap",
+    "build": "geoprocessing build:lambda && geoprocessing build:client",
+    "build:client": "geoprocessing build:client",
+    "build:lambda": "geoprocessing build:lambda",
+    "clear-all-results": "geoprocessing clear-all-results",
+    "clear-results": "geoprocessing clear-results",
+    "create:client": "geoprocessing create:client",
+    "create:function": "geoprocessing create:function",
+    "create:report": "geoprocessing create:report",
+    "deploy": "geoprocessing deploy",
+    "destroy": "geoprocessing destroy",
+    "import:data": "geoprocessing import:data",
+    "precalc:data": "start-server-and-test 'http-server data/dist -c-1 -p 8001' http://localhost:8001 precalc:data_",
+    "precalc:data:clean": "geoprocessing precalc:data:clean",
+    "precalc:data_": "geoprocessing precalc:data",
+    "publish:data": "geoprocessing publish:data",
+    "reimport:data": "geoprocessing reimport:data",
+    "start-data": "http-server data/dist -c-1",
+    "start:client": "geoprocessing start:client",
+    "storybook": "PROJECT_PATH=$(pwd) npx tsx .storybook/genReportStories.ts && storybook dev -p 6006",
+    "storybook:install": "npx tsx scripts/storybookInstall.ts",
+    "synth": "geoprocessing synth",
+    "test": "start-server-and-test start-data 8080 'npm run __test'",
+    "test:matching": "npm run __test -- -t",
+    "translation:extract": "npx tsx scripts/translationExtract.ts",
+    "translation:import": "npx tsx src/i18n/bin/importTerms.ts",
+    "translation:publish": "npx tsx src/i18n/bin/publishTerms.ts",
+    "translation:sync": "npm run translation:extract && npm run translation:publish && npm run translation:import",
+    "upgrade": "geoprocessing upgrade",
+    "url": "geoprocessing url"
   },
-  "homepage": "https://github.com/seasketch/fsm-nearshore-reports#readme",
-  "bugs": {
-    "url": "https://github.com/seasketch/fsm-nearshore-reports/issues"
-  }
+  "private": false,
+  "type": "module",
+  "keywords": [
+    "@seasketch/geoprocessing"
+  ],
+  "repositoryUrl": "https://github.com/seasketch/fsm-nearshore-reports"
 }
