+++ project/i18n.json
@@ -1,0 +1,4 @@
+{
+  "localNamespace": "translation",
+  "remoteContext": "fsm-nearshore-worker"
+}
