+++ src/i18n/bin/importTerms.ts
@@ -2,7 +2,6 @@
 /* eslint-disable i18next/no-literal-string */
 import fs from "fs-extra";
 import * as path from "path";
-import config from "../config.json";
 import { Translations } from "./publishTerms.js";
 
 if (!process.env.POEDITOR_API_TOKEN) {
@@ -11,6 +10,24 @@ if (!process.env.POEDITOR_API_TOKEN) {
 if (!process.env.POEDITOR_PROJECT) {
   throw new Error("POEDITOR_PROJECT is not defined");
 }
+
+const installedProjectPath = path.join("project"); // assumes script running from top-level
+const monoProjectPath = path.join("packages/geoprocessing/src/project");
+
+const projectPath = (() => {
+  if (fs.existsSync(installedProjectPath)) {
+    console.log(`Found project path at ${installedProjectPath}`);
+    return installedProjectPath;
+  } else if (fs.existsSync(monoProjectPath)) {
+    console.log(`Found project path at ${monoProjectPath}`);
+    return monoProjectPath;
+  }
+  throw new Error(
+    `Could not find path to project dir, tried ${installedProjectPath} and ${monoProjectPath}`
+  );
+})();
+
+const config = await fs.readJSON(`${projectPath}/i18n.json`);
 
 /**
  * Pulls all terms for all languages from POEditor and writes them to the local translation files.
