+++ src/i18n/extraTerms.json
@@ -1,4 +1,51 @@
 {
   "Description": "Description",
-  "Author(s)": "Author(s)"
+  "Author(s)": "Author(s)",
+  "Micronesian Exclusive Economic Zone": "Micronesian Exclusive Economic Zone",
+  "Marine Protected Areas": "Marine Protected Areas",
+  "Aquaculture Areas": "Aquaculture Areas",
+  "Territorial Sea": "Territorial Sea",
+  "Aquaculture": "Aquaculture",
+  "Community/Recreational": "Community/Recreational",
+  "Construction/Infrastructure": "Construction/Infrastructure",
+  "Cultural Use": "Cultural Use",
+  "Fisheries": "Fisheries",
+  "Maritime Activity": "Maritime Activity",
+  "Research": "Research",
+  "Tourism": "Tourism",
+  "Transportation": "Transportation",
+  "Coral/Algae": "Coral/Algae",
+  "Microalgal Mats": "Microalgal Mats",
+  "Rock": "Rock",
+  "Rubble": "Rubble",
+  "Sand": "Sand",
+  "Seagrass": "Seagrass",
+  "Nearshore": "Nearshore",
+  "Key Benthic Habitat": "Key Benthic Habitat",
+  "Total": "Total",
+  "Community Recreational Use": "Community Recreational Use",
+  "Research & Data Collection": "Research & Data Collection",
+  "Construction & Infrastructure": "Construction & Infrastructure",
+  "Lelu": "Lelu",
+  "Malem": "Malem",
+  "Tafunsak": "Tafunsak",
+  "Utwe": "Utwe",
+  "Walung": "Walung",
+  "Unknown": "Unknown",
+  "Bottom Fishing": "Bottom Fishing",
+  "Cast Net": "Cast Net",
+  "Casting": "Casting",
+  "Gathering & Reef Gleaning": "Gathering & Reef Gleaning",
+  "Other (Torch Fishing)": "Other (Torch Fishing)",
+  "Spear & Harpoon": "Spear & Harpoon",
+  "Trap, Fish Fence, & Gill Net": "Trap, Fish Fence, & Gill Net",
+  "Trolling": "Trolling",
+  "Vertical Longline": "Vertical Longline",
+  "Coastal Objective": "Coastal Objective",
+  "mpa": "mpa",
+  "aquaculture": "aquaculture",
+  "Chuuk": "Chuuk",
+  "Kosrae": "Kosrae",
+  "Pohnpei": "Pohnpei",
+  "Yap": "Yap"
 }
